Rename schema.py to bigquery_schema.py
